style: make widened multiplications explicit
bugprone-implicit-widening-of-multiplication-result across 24 files: an int
multiplication feeding a size_t parameter now says so at the call site. All of
the affected expressions are constant and none can overflow; the value is in
stating the intent rather than in fixing a latent bug.

Three separate defects in clang-tidy's own fixits had to be repaired by hand:

- it emitted `static_cast<size_type>`, an unqualified nested typedef that is not
  in scope at those call sites (24 occurrences, now `std::size_t`);
- it dropped the closing parenthesis when the expression sat inside a Catch2
  CHECK/REQUIRE macro, leaving 15 unbalanced statements;
- header fixes were applied once per including translation unit, producing
  doubly-nested `static_cast<size_t const>(static_cast<size_t const>(...))`.

readability-convert-member-functions-to-static was attempted alongside this and
is left disabled, as it was before. Its fix places `static` ahead of
`[[nodiscard]]`, which makes the attribute appertain to the return type, and it
turns Qt Q_INVOKABLE members into static functions while dropping their const --
a change to what QML binds to, not a style fix.

+++ src/contour/Config.h
@@ -48,6 +48,7 @@
 
 #include <chrono>
 #include <concepts>
+#include <cstddef>
 #include <cstdint>
 #include <exception>
 #include <expected>
@@ -1928,14 +1929,16 @@ struct YAMLConfigWriter: Writer
     template <typename... T>
     std::string process(std::string_view doc, T... val)
     {
-        return format(addOffset(replaceCommentPlaceholder(std::string { doc }), Offset::Levels * OneOffset),
+        return format(addOffset(replaceCommentPlaceholder(std::string { doc }),
+                                static_cast<size_t>(Offset::Levels * OneOffset)),
                       val...);
     }
 
     template <typename... T>
     std::string process(std::string_view doc, [[maybe_unused]] std::string_view name, T... val)
     {
-        return format(addOffset(replaceCommentPlaceholder(std::string { doc }), Offset::Levels * OneOffset),
+        return format(addOffset(replaceCommentPlaceholder(std::string { doc }),
+                                static_cast<size_t>(Offset::Levels * OneOffset)),
                       val...);
     }
