+++ src/vtbackend/MessageParser.h
@@ -5,6 +5,7 @@
 
 #include <crispy/algorithm.h>
 
+#include <cstddef>
 #include <cstdint>
 #include <string>
 #include <string_view>
@@ -88,7 +89,7 @@ class MessageParser: public ParserExtension
     constexpr static size_t MaxKeyLength = 64;
     constexpr static size_t MaxValueLength = 512;
     constexpr static size_t MaxParamCount = 32;
-    constexpr static size_t MaxBodyLength = 16 * 1024 * 1024; // 16 MB
+    constexpr static size_t MaxBodyLength = static_cast<size_t>(16 * 1024 * 1024); // 16 MB
 
     using OnFinalize = std::function<void(Message&&)>;
